+++ src/main/java/com/neo/web/UserInfoController.java
@@ -62,7 +62,9 @@ public class UserInfoController {
 		if(userId==null){
 			return new MessageResult(-1,"参数错误");
 		}
-		
+		if("admin".equals(userId)){
+			return new MessageResult(-1,"超级管员不可删除");
+		}
 		return userInfoService.deleteUInfoById(userId);
 	}
